+++ lib/rixie/llm/adapter/openai.rb
@@ -10,11 +10,11 @@ module Rixie
   module LLM
     module Adapter
       class OpenAI
-        def initialize(model:, base_url:, api_key:, request_timeout: nil, max_tokens: nil, temperature: nil, parallel_tool_calls: true)
+        def initialize(model:, base_url:, api_key:, request_timeout: nil, temperature: nil, parallel_tool_calls: true, provider_params: nil)
           @model = model
-          @max_tokens = max_tokens
           @temperature = temperature
           @parallel_tool_calls = parallel_tool_calls
+          @provider_params = provider_params || {}
           params = {api_key: api_key, base_url: base_url}
           params[:timeout] = request_timeout if request_timeout
           @client = ::OpenAI::Client.new(**params)
@@ -117,8 +117,8 @@ module Rixie
             params[:tools] = encode_tools(tools)
             params[:parallel_tool_calls] = @parallel_tool_calls
           end
-          params[:max_tokens] = @max_tokens if @max_tokens
           params[:temperature] = @temperature unless @temperature.nil?
+          params.merge!(@provider_params)
           params
         end